move server files into lib folder, modularise routes
