+++ lib/redisFunctions.js
