+++ lib/routes/addPost.js
@@ -1,0 +1,9 @@
+const redisFunctions = require('../redisFunctions.js');
+module.exports = {
+  method: 'POST',
+  path: '/admindotfilter',
+  handler: (request, reply) => {
+    redisFunctions.addPostToDB(request.payload);
+    reply.redirect('/blog');
+  }
+};
