+++ lib/routes/resources.js
@@ -1,0 +1,7 @@
+module.exports = {
+  method: 'GET',
+  path: '/{param*}',
+  handler: {
+    directory: { path: 'public' }
+  }
+};
