+++ public/scss/_variables.scss
@@ -1,2 +1,2 @@
-$font: Roboto Condensed;
+$font: Roboto Condensed, Roboto, Lato, Helvetica;
 $yellow: #F9C80D;
