+++ tests/auth.test.js
@@ -1,9 +1,9 @@
 "use strict";
 
 const tape = require('tape');
-const redisFunctions = require('../redisFunctions.js');
-const server = require('../server.js');
-const auth = require('../auth.js');
+const redisFunctions = require('../lib/redisFunctions.js');
+const server = require('../lib/server.js');
+const auth = require('../lib/auth.js');
 const client = redisFunctions.client;
 
 let basicHeader = (username, password) => {
