+++ views/home.html
@@ -19,6 +19,6 @@
       {{>posts this}}
     {{/each}}
   </div>
-  {{>profiles}}
+  <!-- {{>profiles}} -->
 </div>
 {{>footer}}
